refactor: event 패키지로 분류
