+++ app/src/main/java/woowacourse/shopping/view/cart/event/CartAdapterEventHandler.kt
@@ -1,4 +1,4 @@
-package woowacourse.shopping.view.cart.adatper
+package woowacourse.shopping.view.cart.event
 
 fun interface CartAdapterEventHandler {
     fun onClickDeleteItem(id: Long)
