+++ app/src/main/java/woowacourse/shopping/view/detail/DetailActivity.kt
@@ -14,6 +14,7 @@ import androidx.databinding.DataBindingUtil
 import woowacourse.shopping.R
 import woowacourse.shopping.databinding.ActivityDetailBinding
 import woowacourse.shopping.view.cart.CartActivity
+import woowacourse.shopping.view.detail.event.DetailScreenEventHandler
 import woowacourse.shopping.view.detail.vm.DetailViewModel
 import woowacourse.shopping.view.detail.vm.DetailViewModelFactory
