+++ app/src/main/java/woowacourse/shopping/view/detail/event/DetailScreenEventHandler.kt
@@ -1,4 +1,4 @@
-package woowacourse.shopping.view.detail
+package woowacourse.shopping.view.detail.event
 
 fun interface DetailScreenEventHandler {
     fun onClickAddToCart()
