+++ app/src/main/java/woowacourse/shopping/view/main/adapter/ProductViewHolder.kt
@@ -3,6 +3,7 @@ package woowacourse.shopping.view.main.adapter
 import androidx.recyclerview.widget.RecyclerView
 import woowacourse.shopping.databinding.ItemProductBinding
 import woowacourse.shopping.domain.Product
+import woowacourse.shopping.view.main.event.ProductsAdapterEventHandler
 
 class ProductViewHolder(
     private val binding: ItemProductBinding,
